requirements.txt
PyQt6>=6.4.0
bcrypt>=4.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0
passlib>=1.7.4